# src/internal/web_api/controllers/short_link_controller.go
package controllers

import (
	"encoding/json"
	"errors"
	"net/http"
	shortlink "shortener/src/internal/domain/short_link"
	"shortener/src/internal/domain/visit"
	"shortener/src/internal/web_api/models"
	"shortener/src/pkg/logger"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/go-playground/validator/v10"
	"github.com/google/uuid"
)

type ShortLinkController struct {
	shortLinkService shortlink.ShortLinkService
	visitService     visit.VisitService
	validator        *validator.Validate
}

func NewShortLinkController(
	shortLinkService shortlink.ShortLinkService,
	visitService visit.VisitService,
	validator *validator.Validate,
) *ShortLinkController {
	return &ShortLinkController{
		shortLinkService: shortLinkService,
		visitService:     visitService,
		validator:        validator,
	}
}

func (c *ShortLinkController) UseHandlers(r chi.Router) {
	r.Post("/shorten", c.Create)
	r.Get("/s/{short_url}", c.Redirect)
}

// Create godoc
//
//	@Summary		Создать короткую ссылку
//	@Description	Создаёт новую короткую ссылку. Если shortCode не задан, он генерируется.
//	@Tags			shortlink
//	@Accept			json
//	@Produce		json
//	@Param			request	body		models.CreateShortLinkRequest	true	"Данные для создания короткой ссылки"
//	@Success		201		{object}	models.CreateShortLinkResponse
//	@Failure		400		{string}	string	"bad request"
//	@Failure		409		{string}	string	"short link already exists"
//	@Failure		500		{string}	string	"internal error"
//	@Router			/shorten [post]
func (c *ShortLinkController) Create(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	var req models.CreateShortLinkRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	if err := c.validator.StructCtx(ctx, req); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	shortLink, err := c.shortLinkService.Create(ctx, req.ShortURLString(), req.OriginalURL)
	if err != nil {
		if errors.Is(err, shortlink.ErrShortLinkAlreadyExists) {
			http.Error(w, err.Error(), http.StatusConflict)
			return
		}

		logger.Error("failed to create short link", "err", err)
		http.Error(w, err.Error(), http.StatusInternalServerError)
	}

	res := models.ShortLinkToCreateResponse(*shortLink)

	w.WriteHeader(http.StatusCreated)
	if err := json.NewEncoder(w).Encode(res); err != nil {
		logger.Error("failed to write response", "err", err)
	}
}

// Redirect godoc
//
//	@Summary		Перенаправить по короткой ссылке
//	@Description	Ищет короткую ссылку, регистрирует визит и перенаправляет на исходный URL.
//	@Tags			shortlink
//	@Param			short_url	path	string	true	"Короткий код"
//	@Success		302			"Redirect"
//	@Failure		404			{string}	string	"short link not found"
//	@Failure		500			{string}	string	"internal error"
//	@Router			/s/{short_url} [get]
func (c *ShortLinkController) Redirect(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	shortURL := chi.URLParam(r, "short_url")

	shortLink, err := c.shortLinkService.Get(ctx, shortURL)
	if err != nil {
		if errors.Is(err, shortlink.ErrShortLinkNotFound) {
			http.Error(w, err.Error(), http.StatusNotFound)
			return
		}

		logger.Error("failed to get short link", "err", err)
		http.Error(w, err.Error(), http.StatusInternalServerError)
	}

	visit := visit.Visit{
		ID:        uuid.New(),
		LinkID:    shortLink.ID,
		CreatedAt: time.Now(),
		UserAgent: r.UserAgent(),
		IPAddress: r.RemoteAddr,
	}

	if err := c.visitService.Register(ctx, visit); err != nil {
		logger.Error("failed to register visit", "err", err)
	}

	http.Redirect(w, r, shortLink.OriginalURL, http.StatusFound)
}