# src/internal/web_api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис для создания коротких ссылок и получения аналитики по переходам.",
        "title": "Shortener API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/analytics/{short_url}": {
            "get": {
                "description": "Возвращает статистику переходов, агрегированную по дням, месяцам или User-Agent.",
                "tags": [
                    "analytics"
                ],
                "summary": "Получить аналитику по короткой ссылке",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Короткий код",
                        "name": "short_url",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "day",
                            "month",
                            "userAgent"
                        ],
                        "type": "string",
                        "description": "Тип группировки",
                        "name": "group",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результат зависит от типа группировки",
                        "schema": {}
                    },
                    "400": {
                        "description": "unknown group",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/s/{short_url}": {
            "get": {
                "description": "Ищет короткую ссылку, регистрирует визит и перенаправляет на исходный URL.",
                "tags": [
                    "shortlink"
                ],
                "summary": "Перенаправить по короткой ссылке",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Короткий код",
                        "name": "short_url",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect"
                    },
                    "404": {
                        "description": "short link not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shorten": {
            "post": {
                "description": "Создаёт новую короткую ссылку. Если shortCode не задан, он генерируется.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shortlink"
                ],
                "summary": "Создать короткую ссылку",
                "parameters": [
                    {
                        "description": "Данные для создания короткой ссылки",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateShortLinkRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CreateShortLinkResponse"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "short link already exists",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CreateShortLinkRequest": {
            "type": "object",
            "required": [
                "originalURL"
            ],
            "properties": {
                "originalURL": {
                    "type": "string"
                },
                "shortURL": {
                    "type": "string",
                    "maxLength": 6
                }
            }
        },
        "models.CreateShortLinkResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "originalURL": {
                    "type": "string"
                },
                "shortCode": {
                    "type": "string"
                }
            }
        }
    }
}